Add files via upload
Build a house on each B site. A house is made of two cylinders and a cone roof. The rover can only carry one building piece at a time with an electromagnet. This code finds the shortest path to build the three houses by creating a 3D array of binary decision variables. The dimensions of the array are start position, end position, and turn. The number of moves needed is deterministic in this scenario because blocks must go directly from their home location to a building site and are not to be left on the ground. There are eight points of interest on the map and each of the six blocks will take two turns to move. The array of decision variables is 8x8x12 to build two simple houses. Although this would not scale well, It works for the project at hand. For a larger project, a well designed recursive algorithm would easily out preform this code.